// src/app/core/Models/distribution/idistribution.ts
export interface Idistribution {
  id?: number;
  client_name: string;
  client_phone: string;
  source_company: string;
  client_category: string;
  client_location: string;
  entry_channe: string;
  entry_campaign: string;
  gender: string;
  request_type: string;
  client_main_domain: string;
  client_sub_domain: string;
  region: string;
  createdAt: string;
}